database/requirements.txt
# Database Migration Requirements
alembic==1.13.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0

